Rename swagger to openapi; endpoint to /docs
