--- swagger/template.html
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>{{.Title}}</title>
-    <meta charset="utf-8"/>
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link href="https://fonts.googleapis.com/css?family=Montserrat:300,400,700|Roboto:300,400,700" rel="stylesheet">
-    <style>
-      body {
-        margin: 0;
-        padding: 0;
-      }
-    </style>
-  </head>
-  <body>
-    <redoc spec-url='{{.Spec}}'></redoc>
-    <script src="https://cdn.jsdelivr.net/npm/redoc@2.0.0-alpha.20/bundles/redoc.standalone.js"> </script>
-  </body>
-</html>
